% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.R
\name{AssemPar}
\alias{AssemPar}
\title{Estimate parameters of the assemblage.}
\usage{
AssemPar(data, type, spatial.str = FALSE, coords, moranI.alfa = 0.05,
  spatial.unsee.sp = "positive")
}
\arguments{
\item{data}{Dataframe giving P/A or abundance (counts or coverage) of species.}

\item{type}{Nature of the data to be processed. It may be presence / absence
("P/A"), counts of individuals ("counts"), or coverage ("cover").}

\item{spatial.str}{Logical argument indicating whether spatial information
(coordinates) of samples is available. If set TRUE, a matrix or dataframe
with X and Y coordinates must be supplied. This file must have the same
number of rows than data to be used by the function Moran.I{ape}. The
analisys is only valid for individual counts (argument type = "counts"), and
will gave an Error if two or more samples share the same X and Y values.}

\item{coords}{Matrix or dataframe given Y and X coordinates of samples.}

\item{moranI.alfa}{Probability for rejecting random spatial distribution. Default is < 0.05.}

\item{spatial.unsee.sp}{Spatial distribution to be assumed for unseen species. It might be "random", "positive" or "negative".}
}
\value{
The function returns an object of class List, to be used by SimDat.
}
\description{
Assempar is a function that extracts the main parameters of the pilot data using basic R functions as well as functions like specpool{Vegan} and Moran.I{ape}.
}
\examples{
AssemPar(data, type, spatial.str, coords, moranI.alfa, spatial.unsee.sp)
}
