\name{ioptimum}
\alias{ioptimum}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Identification of the optimal sampling effort
}
\description{
The function estimates the sampling effort in which the rate of change for each additional sampling unit can be considered optimal.
}
\usage{
ioptimum(xx, multi.site = TRUE, c1 = 10, c2 =5, c3= 1)
}

\arguments{
  \item{xx}{
A data frame generated by \code{\link{summary_ssp}}
}
\item{multi.site}{
Logical argument indicating if several sites were simulated
  }
\item{c1}{First cut. By default 10\% improvement for each sample with respect to the highest MultSE. }
\item{c2}{Second cut. By default 5\% improvement for each sample with respect to the highest MultSE. }
\item{c3}{Third cut. By default 1\% improvement for each sample with respect to the highest MultSE.}
}
\details{
Between the minimum sampling effort (i.e. 2) and c1, can be considered the necessary improvement for each aditional sample. The sampling efforts between c1 and c2 would indicate the optimal number of samples. A cost / benefit criterion can be used to set the final sample size. The sampling range between c2 and c3 would imply a marginal improvement of the MultSE for each increase in sample size, which would result in an unnecessary sampling effort due to redundancy.
}
\value{
\item{sample.cut}{
A matrix with the sampling size definig the range of insufficient, optimal and redundant effort
}
}
\references{
Underwood, A. J. (1990). Experiments in ecology and management: Their logics, functions and interpretations. Australian Journal of Ecology, 15, 365-389.
}
\author{
Edlin Guerra-Castro (edlinguerra@gmail.com), Maite Mascaro (mmm@ciencias.unam.mx), Juan Carlos Cajas (carlos.cajas@enesmerida.unam.mx)
}
\note{
The cuts that define the sampling effort as necessary, optimal or redundant are arbitrary and can be modified according to each research problem. In particular, it is possible that c3 as 1\% is not generated because this would be achieved with a sample size larger than the maximun simulated. In this case, c3 can be set as 2\% or 3\%.
}

\seealso{
\code{\link[SSP]{sampsd}}
}
\examples{
###To speed up the simulation of these examples, the cases, sites and n were set small.

##Single site: micromollusk from Cayo Nuevo (Yucatan, Mexico)
data(micromollusk)

#Estimation of parameters of pilot data
par.mic<-assempar (data = micromollusk,
                    type= "P/A",
                    Sest.method = "average")

#Simulation of 10 data sets, each one with 100 potential sampling units from a single site
sim.mic<-simdata(par.mic, cases= 10, n= 100, sites = 1)

#Sampling and estimation of MultSE for each sample size (few repetitions to speed up the example)
sam.mic<-sampsd(sim.mic,
               par.mic,
               transformation = "P/A",
               method = "jaccard",
               multi.site = FALSE,
               n=100,
               p.n = 50,
               sites = 1,
               p.s = 1,
               k=3)

#Summary of MultSE for each sampling effort
summ.mic<-summary_ssp(results = sam.mic, multi.site = FALSE)

opt.mic<-ioptimum(summ.mic, multi.site = FALSE, c1 = 10, c2 =5, c3= 1)

##Multiple sites: Sponges from Alacranes National Park (Yucatan, Mexico).
data(sponges)

#Estimation of parameters of pilot data
par.spo<-assempar(data = sponges,
                  type= "counts",
                  Sest.method = "average")

#Simulation of 3 data sets, each one with 15 potential sampling units in 10 sites.
sim.spo<-simdata(par.spo, cases= 10, n= 15, sites = 10)

#Sampling and estimation of MultSE for each sampling design (few
#repetitions to speed up the example)

sam.spo<-sampsd(sim.spo,
                par.spo,
                transformation = "square root",
                method = "bray",
                multi.site = TRUE,
                n=15,
                p.n = 10,
                sites = 10,
                p.s = 10,
                k=3)

#Summary of MultSE for each sampling effort
summ.spo<-summary_ssp(results = sam.spo, multi.site = TRUE)

opt.spo<-ioptimum(summ.spo, multi.site = TRUE, c1 = 10, c2 =5, c3= 1)

}

\keyword{ multivariate	&	Multivariate Techniques }
\keyword{ ecological communities }
\keyword{ numerical ecology }
