\name{ioptsd}
\alias{ioptsd}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Identification of the optimal sampling design
}
\description{
Regions of sampling effort where the rate of change for each aditional sampling unit is less than or equal to 10\%, 5\% and 1\% of the highest MultSE.
}
\usage{
ioptsd(mse.summary, multi.site = TRUE, c1 = 10, c2 =5, c3= 1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{mse.summary}{
A data frame generated by \code{\link{summary.ssp}}
}
\item{multi.site}{
Logical argument indicating if several sites were simulated
  }
\item{c1}{First cut. By default 10\% improvement for each sample with respect to the highest MultSE. Between the minimum sampling effort and this cut, it can be considered insufficient improvement.}
\item{c2}{Second cut. By default 5\% improvement for each sample with respect to the highest MultSE. The sampling range between c1 and c2 would indicate the optimal number of samples based on a cost / benefit criterion}
\item{c3}{Third cut. By default 1\% improvement for each sample with respect to the highest MultSE. The sampling range between c2 and c1 would imply a marginal improvement of the MultSE, which would result in an unnecessary sampling effort due to redundancy}
}
\details{
Regions in the sampling effort where the rate of change of the relativized average MultSE is less than or equal to the 10\% (insufficient), 5\% (optimal) and 1\% (redundant) are identified. The values are summarized in a data frame.
}
\value{
%%  ~Describe the value returned
%%  If it is a LIST, use
%%  \item{comp1 }{Description of 'comp1'}
%%  \item{comp2 }{Description of 'comp2'}
%% ...
}
\references{
%% ~put references to the literature/web site here ~
}
\author{
%%  ~~who you are~~
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--	or do  help(data=index)  for the standard data sets.

## The function is currently defined as
function (x)
{
  }
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ ~kwd1 }% use one of  RShowDoc("KEYWORDS")
\keyword{ ~kwd2 }% __ONLY ONE__ keyword per line
