\name{micromollusk}
\alias{micromollusk}
\docType{data}
\title{
Micromollusks of marine shallow sandy bottoms around Cayo Nuevo, Gulf of Mexico, Mexico
}
\description{Presence/absence of 67 species registered in six cores of 4-inch diameter and 10 cm depth taken in sandy bottoms around Cayo Nuevo, Gulf of Mexico, Mexico
}
\usage{data("micromollusk")}
\format{
  A data frame with 6 observations on the following 68 variables.
  \describe{
    \item{\code{Leptochiton.sp.}}{a numeric vector}
    \item{\code{Ischnochiton..Ischnochiton..erythronotus}}{a numeric vector}
    \item{\code{Arcidae.sp.}}{a numeric vector}
    \item{\code{Arca.imbricata}}{a numeric vector}
    \item{\code{Barbatia.domingensis}}{a numeric vector}
    \item{\code{Bentharca.sp.}}{a numeric vector}
    \item{\code{Arcopsis.adamsi}}{a numeric vector}
    \item{\code{Crenella.sp.}}{a numeric vector}
    \item{\code{Anomia.sp..}}{a numeric vector}
    \item{\code{Carditopsis.smithii}}{a numeric vector}
    \item{\code{Lucinidae..}}{a numeric vector}
    \item{\code{Chama.sinuosa}}{a numeric vector}
    \item{\code{Chama.sp.}}{a numeric vector}
    \item{\code{Galeommatidae.sp.}}{a numeric vector}
    \item{\code{Chione.elevata}}{a numeric vector}
    \item{\code{Semele.bellastriata}}{a numeric vector}
    \item{\code{Gastropoda.sp..1..}}{a numeric vector}
    \item{\code{Gastropoda.sp..2..}}{a numeric vector}
    \item{\code{Gastropoda.sp..3..}}{a numeric vector}
    \item{\code{Diodora.minuta}}{a numeric vector}
    \item{\code{Diodora.sp...}}{a numeric vector}
    \item{\code{Scissurella.redferni}}{a numeric vector}
    \item{\code{Synaptocochlea.picta}}{a numeric vector}
    \item{\code{Lodderena.ornata}}{a numeric vector}
    \item{\code{Cerithium.sp...}}{a numeric vector}
    \item{\code{Sansonia.tuberculata}}{a numeric vector}
    \item{\code{Iniforis.turristhomae}}{a numeric vector}
    \item{\code{Metaxia.rugulosa}}{a numeric vector}
    \item{\code{Cerithiopsis.cf..iuxtafuniculata}}{a numeric vector}
    \item{\code{Cerithiopsis.sp.}}{a numeric vector}
    \item{\code{Vermetidae.incertae.sedis.irregularis}}{a numeric vector}
    \item{\code{Dendropoma.corrodens}}{a numeric vector}
    \item{\code{Vermetid.sp..C}}{a numeric vector}
    \item{\code{Petaloconchus.mcgintyi}}{a numeric vector}
    \item{\code{Thylacodes.sp.}}{a numeric vector}
    \item{\code{Alvania.auberiana}}{a numeric vector}
    \item{\code{Alvania.colombiana}}{a numeric vector}
    \item{\code{Alvania.sp.}}{a numeric vector}
    \item{\code{Simulamerelina.caribaea}}{a numeric vector}
    \item{\code{Schwartziella.fischeri}}{a numeric vector}
    \item{\code{Zebina.browniana}}{a numeric vector}
    \item{\code{Zebina.sp.}}{a numeric vector}
    \item{\code{Caecum.circumvolutum}}{a numeric vector}
    \item{\code{Caecum.donmoorei}}{a numeric vector}
    \item{\code{Caecum.floridanum}}{a numeric vector}
    \item{\code{Caecum.johnsoni}}{a numeric vector}
    \item{\code{Caecum.pulchellum}}{a numeric vector}
    \item{\code{Caecum.textile}}{a numeric vector}
    \item{\code{Caecum.sp..B}}{a numeric vector}
    \item{\code{Meioceras.nitidum}}{a numeric vector}
    \item{\code{Cochliolepis.striata}}{a numeric vector}
    \item{\code{Parviturboides.interruptus}}{a numeric vector}
    \item{\code{Vitrinella.sp.}}{a numeric vector}
    \item{\code{Gibberula.lavalleeana}}{a numeric vector}
    \item{\code{Prunum.apicinum}}{a numeric vector}
    \item{\code{Volvarina.avena}}{a numeric vector}
    \item{\code{Astyris.lunata}}{a numeric vector}
    \item{\code{Phrontis.albus}}{a numeric vector}
    \item{\code{Phrontis.sp.}}{a numeric vector}
    \item{\code{Trachypollia.sp...}}{a numeric vector}
    \item{\code{Turridae.sp..1}}{a numeric vector}
    \item{\code{Turridae.sp..2..}}{a numeric vector}
    \item{\code{Turridae.sp..3..}}{a numeric vector}
    \item{\code{Ammonicera.lineofuscata}}{a numeric vector}
    \item{\code{Ammonicera.minortalis}}{a numeric vector}
    \item{\code{Rissoella.galba}}{a numeric vector}
    \item{\code{Pyramidellidae.sp.}}{a numeric vector}
    \item{\code{Pseudoscilla.babylonia}}{a numeric vector}
  }
}
\details{

cayo Nuevo is a small reef cay located 240 km off the North-Western coast of Yucatan. Data correspond to a study on the biodiversity of marine benthic reef habitats off the Yucatan shelf (Ortigosa, Suarez-Mozo, Barrera et al. 2018).
}
\source{
https://doi.org/10.3897/zookeys.779.24562
}
\references{
Ortigosa, D., Suarez-Mozo, N. Y., Barrera, N. C., & Simoes, N. (2018). First survey of Interstitial molluscs from Cayo Nuevo, Campeche Bank, Gulf of Mexico. Zookeys, 779. doi:10.3897/zookeys.779.24562
}
\examples{
data(micromollusk)

}
\keyword{datasets}
