\name{micromolluskGEO}
\alias{micromolluskGEO}
\docType{data}
\title{
Geographic coordinates of cores in \code{\link[SSP]{micromollusk}} data
}

\description{
Spatial coordinates of each core sample
}
\usage{data("micromolluskGEO")}
\format{
  A data frame with 6 observations on the following 2 variables.
  \describe{
    \item{\code{Lat}}{a numeric vector}
    \item{\code{Lon}}{a numeric vector}
  }
}
\details{
Geographic coordinates are given in decimal degrees
}
\source{
https://doi.org/10.3897/zookeys.779.24562
}
\references{
Ortigosa, D., Suarez-Mozo, N. Y., Barrera, N. C., & Simoes, N. (2018). First survey of Interstitial molluscs from Cayo Nuevo, Campeche Bank, Gulf of Mexico. Zookeys, 779. doi:10.3897/zookeys.779.24562

}
\examples{
data(micromolluskGEO)
}
\keyword{datasets}
