\name{pilot}
\alias{pilot}
\docType{data}
\title{
Epibionts on Caribbean mangrove roots: pilot data
}
\description{
Data corresponds to a pilot study abput epibenthic organisms in mangrove roots from Laguna de La Restinga National Park, Venezuela (Guerra-Castro et al. 2011).
}
\usage{data("pilot")}
\format{
  A data frame with 180 observations on the following 118 variables.
  \describe{
    \item{\code{Sector}}{a factor with levels \code{E} \code{I} \code{M}}
    \item{\code{Site}}{a numeric vector}
    \item{\code{sp1}}{a numeric vector}
    \item{\code{sp2}}{a numeric vector}
    \item{\code{sp3}}{a numeric vector}
    \item{\code{sp4}}{a numeric vector}
    \item{\code{sp5}}{a numeric vector}
    \item{\code{sp6}}{a numeric vector}
    \item{\code{sp7}}{a numeric vector}
    \item{\code{sp8}}{a numeric vector}
    \item{\code{sp9}}{a numeric vector}
    \item{\code{sp10}}{a numeric vector}
    \item{\code{sp11}}{a numeric vector}
    \item{\code{sp12}}{a numeric vector}
    \item{\code{sp13}}{a numeric vector}
    \item{\code{sp14}}{a numeric vector}
    \item{\code{sp15}}{a numeric vector}
    \item{\code{sp16}}{a numeric vector}
    \item{\code{sp17}}{a numeric vector}
    \item{\code{sp18}}{a numeric vector}
    \item{\code{sp19}}{a numeric vector}
    \item{\code{sp20}}{a numeric vector}
    \item{\code{sp21}}{a numeric vector}
    \item{\code{sp22}}{a numeric vector}
    \item{\code{sp23}}{a numeric vector}
    \item{\code{sp24}}{a numeric vector}
    \item{\code{sp25}}{a numeric vector}
    \item{\code{sp26}}{a numeric vector}
    \item{\code{sp27}}{a numeric vector}
    \item{\code{sp28}}{a numeric vector}
    \item{\code{sp29}}{a numeric vector}
    \item{\code{sp30}}{a numeric vector}
    \item{\code{sp31}}{a numeric vector}
    \item{\code{sp32}}{a numeric vector}
    \item{\code{sp33}}{a numeric vector}
    \item{\code{sp34}}{a numeric vector}
    \item{\code{sp35}}{a numeric vector}
    \item{\code{sp36}}{a numeric vector}
    \item{\code{sp37}}{a numeric vector}
    \item{\code{sp38}}{a numeric vector}
    \item{\code{sp39}}{a numeric vector}
    \item{\code{sp40}}{a numeric vector}
    \item{\code{sp41}}{a numeric vector}
    \item{\code{sp42}}{a numeric vector}
    \item{\code{sp43}}{a numeric vector}
    \item{\code{sp44}}{a numeric vector}
    \item{\code{sp45}}{a numeric vector}
    \item{\code{sp46}}{a numeric vector}
    \item{\code{sp47}}{a numeric vector}
    \item{\code{sp48}}{a numeric vector}
    \item{\code{sp49}}{a numeric vector}
    \item{\code{sp50}}{a numeric vector}
    \item{\code{sp51}}{a numeric vector}
    \item{\code{sp52}}{a numeric vector}
    \item{\code{sp53}}{a numeric vector}
    \item{\code{sp54}}{a numeric vector}
    \item{\code{sp55}}{a numeric vector}
    \item{\code{sp56}}{a numeric vector}
    \item{\code{sp57}}{a numeric vector}
    \item{\code{sp58}}{a numeric vector}
    \item{\code{sp59}}{a numeric vector}
    \item{\code{sp60}}{a numeric vector}
    \item{\code{sp61}}{a numeric vector}
    \item{\code{sp62}}{a numeric vector}
    \item{\code{sp63}}{a numeric vector}
    \item{\code{sp64}}{a numeric vector}
    \item{\code{sp65}}{a numeric vector}
    \item{\code{sp66}}{a numeric vector}
    \item{\code{sp67}}{a numeric vector}
    \item{\code{sp68}}{a numeric vector}
    \item{\code{sp69}}{a numeric vector}
    \item{\code{sp70}}{a numeric vector}
    \item{\code{sp71}}{a numeric vector}
    \item{\code{sp72}}{a numeric vector}
    \item{\code{sp73}}{a numeric vector}
    \item{\code{sp74}}{a numeric vector}
    \item{\code{sp75}}{a numeric vector}
    \item{\code{sp76}}{a numeric vector}
    \item{\code{sp77}}{a numeric vector}
    \item{\code{sp78}}{a numeric vector}
    \item{\code{sp79}}{a numeric vector}
    \item{\code{sp80}}{a numeric vector}
    \item{\code{sp81}}{a numeric vector}
    \item{\code{sp82}}{a numeric vector}
    \item{\code{sp83}}{a numeric vector}
    \item{\code{sp84}}{a numeric vector}
    \item{\code{sp85}}{a numeric vector}
    \item{\code{sp86}}{a numeric vector}
    \item{\code{sp87}}{a numeric vector}
    \item{\code{sp88}}{a numeric vector}
    \item{\code{sp89}}{a numeric vector}
    \item{\code{sp90}}{a numeric vector}
    \item{\code{sp91}}{a numeric vector}
    \item{\code{sp92}}{a numeric vector}
    \item{\code{sp93}}{a numeric vector}
    \item{\code{sp94}}{a numeric vector}
    \item{\code{sp95}}{a numeric vector}
    \item{\code{sp96}}{a numeric vector}
    \item{\code{sp97}}{a numeric vector}
    \item{\code{sp98}}{a numeric vector}
    \item{\code{sp99}}{a numeric vector}
    \item{\code{sp100}}{a numeric vector}
    \item{\code{sp101}}{a numeric vector}
    \item{\code{sp102}}{a numeric vector}
    \item{\code{sp103}}{a numeric vector}
    \item{\code{sp104}}{a numeric vector}
    \item{\code{sp105}}{a numeric vector}
    \item{\code{sp106}}{a numeric vector}
    \item{\code{sp107}}{a numeric vector}
    \item{\code{sp108}}{a numeric vector}
    \item{\code{sp109}}{a numeric vector}
    \item{\code{sp110}}{a numeric vector}
    \item{\code{sp111}}{a numeric vector}
    \item{\code{sp112}}{a numeric vector}
    \item{\code{sp113}}{a numeric vector}
    \item{\code{sp114}}{a numeric vector}
    \item{\code{sp115}}{a numeric vector}
    \item{\code{sp116}}{a numeric vector}
  }
}
\details{
Data consists of the coverage (by point-intercept) of 116 taxa identified in 180 mangrove roots, sampled under a hierarchically nested spatial design that included six random sites within each of three sectors of the lagoon system corresponding to a strong environmental gradient: external (E), intermediate (M), and internal (I). The abundance of epibenthic organisms of 10 roots were described within each site, producing a total of 60 roots in each sector. The analysis of these pilot data defined the sampling design used by Guerra-Castro et al. (2016).
}
\source{
https://www.interciencia.net/wp-content/uploads/2018/01/923-GUERRA-8.pdf
}
\references{
Guerra-Castro, E., J. J. Cruz-Motta, and J. E. Conde. 2011. Cuantificación de la diversidad de especies incrustantes asociadas a las raíces de Rhizophora mangle L. en el Parque Nacional Laguna de La Restinga. Interciencia 36:923-930.

Guerra-Castro, E. J., J. E. Conde, and J. J. Cruz-Motta. (2016). Scales of spatial variation in tropical benthic assemblages and their ecological relevance: epibionts on Caribbean mangrove roots as a model system. Marine Ecology Progress Series 548:97-110.
}
\examples{
data(pilot)
str(pilot)
}
\keyword{datasets}
\keyword{SSP}
