% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plotMSE}
\alias{plotMSE}
\title{plot MultSE}
\usage{
plotMSE(results, multi.site = TRUE)
}
\arguments{
\item{results}{from prepareMSE}

\item{multi.site}{TRUE or FALSE, default is TRUE}
}
\description{
estimate the average of the MultSE generated for each source of
variation according to the repetitions of each sampling design. Values of the
mean, lower and upper quantiles are tabulated for each sampling effort. The
values are summarized in a dataframe.
}
