% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{prepareMSE}
\alias{prepareMSE}
\title{Prepare data for plotMSE}
\usage{
prepareMSE(results, multi.site = TRUE)
}
\arguments{
\item{results}{dataframe generated by plot.MSE}

\item{multi.site}{Logical argument indicating whether several sites were simulated}
}
\description{
Prepare data for plotMSE
}
