\name{sampsd}
\alias{sampsd}

\title{Sampling simulated data and estimates multivariate standard errors

}
\description{
Each set of simulated data is sampled many times for each sampling effort, from 3 replicates to those defined as an argument in the function. Then, distance-based multivariate standard errors is estimated using pseudo-variance (for single site evaluation) or Mean Squares estimates in a linear model (for multisite evaluation).
}
\usage{
sampsd(dat.sim, Sest, transformation, method, multi.site = TRUE, n, p.n, sites, p.s, k)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{dat.sim}{
A list of data sets generated by \code{\link{simdat}}
}
  \item{Sest}{
Number of species estimated with Chao2. See list generated by \code{\link{assempar}}
}
  \item{transformation}{
Mathematical function to reduce the weight of very dominant species: 'square root', 'fourth root', 'Log (X+1)', 'P/A', 'none'
}
  \item{method}{
The appropiate distance/dissimilarity metric. The function \code{\link[vegan]{vegdist}} is call for that purpose.
  }
  \item{multi.site}{
Logical argument indicating if several sites were simulated
  }
  \item{n}{
Total number of samples simulated for each site
  }
  \item{p.n}{
Maxinum number of samples to take at each site. Can be equal or less than n
  }
  \item{sites}{
Total number of sites simulated for each data set
  }
  \item{p.s}{
Maxinum number of sites to sample at each data set
  }
  \item{k}{
Number of repetitions of each combination between n and sites
  }
}
\details{
Sampling without replacement was performed several times (e.g. 100) for each subset of samples of size n = 3, 4, 5 to N within each simulated matrix. If sites had been generated, subsets of sites of size m = 2, 3, 4 to M were first sampled followed by sampling units for each site using inclusion probabilities and self-weighted two-stage sampling (Tille, 2011). Additionally, each combination of n and m was repeated several times (e.g. 100) for all simulated matrixes. Suitable pre-treatments of data matrixes were applied in each case and distance/similarity matrixes estimated using the appropriate coefficient. When simulations were done for a single site, the MultSE was estimated as \eqn{\sqrtV/n}, being V the pseudo variance measured at each sample of size n (Anderson & Santana-Garcon, 2015). When several sites were generated, MultSE were estimated using the residual mean squares  and the sites mean squares from a PERMANOVA model (Anderson & Santana-Garcon, 2015).
}
\value{
\item{mse.results}{
A data frame indluding all estimated MultSE for each simulated data, combination of sample replicates and sites for each k repetition. This data frame will be used by \code{\link[SSP]{mseplot}}
}
}
\references{
Anderson, M.J. & Santana-Garcon, J. (2015) Measures of precision for dissimilarity-based multivariate analysis of ecological communities. Ecology Letters, 18, 66-73

Tille, Y. (2011) Sampling Algorithms. International Encyclopedia of Statistical Science (ed. M. Lovric), pp. 1273-1274. Springer Berlin Heidelberg, Berlin, Heidelberg
}
\author{
Edlin Guerra-Castro (edlinguerra@gmail.com) and Maite Mascaro (mmm@ciencias.unam.mx)
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## micromollusk from 'Cayo Nuevo', Campeche bank, Gulf of Mexico, Yucatan, Mexico
data(micromollusk)
data(micromolluskGEO)

#Estimation of parameters of pilot data
dat.par1<-assempar (data = micromollusk,
                    type= "P/A",
                    spatial.str = TRUE,
                    coords = micromolluskGEO,
                    moranI.alfa=0.05,
                    spatial.unsee.sp= "random")

#Simulation of 10 data sets, each one with 100 potential sampling units from a single site
dat.sim1<-simdata(dat.par1, cases= 10, n= 100, sites = 1)

#Sampling and estimation of MSE for each sample size
samples1<-sampsd(dat.sim1,
               Sest = 98,
               transformation = "P/A",
               method = "jaccard",
               multi.site = FALSE,
               n=100,
               p.n = 50,
               sites = 1,
               p.s = 1,
               k=10)

##Sponges from Alacranes National Park, Gulf of Mexico
data(sponges)
data(spongesGEO)

#Estimation of parameters of pilot data
dat.par2<-assempar(data = sponges,
                    type= "counts",
                    spatial.str = TRUE,
                    coords = spongesGEO,
                    moranI.alfa=0.05,
                    spatial.unsee.sp= "random")

#Simulation of 10 data sets, each one with 20 potential sampling units (transects) in 20 sites.
dat.sim2<-simdata(dat.par2, cases= 10, n= 20, sites = 20)

#Sampling and estimation of MSE for each combination of sites and samples
samples2<-sampsd(dat.sim2,
                Sest = 50,
                transformation = "square root",
                method = "bray",
                multi.site = TRUE,
                n=20,
                p.n = 10,
                sites = 20,
                p.s = 10,
                k=10)

}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ multivariate	&	Multivariate Techniques }% use one of  RShowDoc("KEYWORDS")
\keyword{ ecological communities }
