\name{simdata}
\alias{simdata}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Simulation of data sets
}
\description{
The function simulates matrixes (as many as requested) using estimated parameters in the list generated by \code{\link{assempar}}. The function returns an object of class List including all the simulated data sets to be used by \code{\link{datquality}} and \code{\link{sampsd}}.
}
\usage{
simdata(Par, cases, n, sites, positive.sp = 0.9, negative.sp= 0.1)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{Par}{
A list of parameters estimated by \code{\link{assempar}}
}
  \item{cases}{
Number of data sets to be simulated
  }
  \item{n}{
Total number of samples to simulate in each site
  }
  \item{sites}{
Total number of sites to simulate in each data set
  }
  \item{positive.sp
  negative.sp}{
Probability of occurrence of a species once it has been detected in a previos sample. The default is 0.9 for positive, 0.1 for negative occurrences
  }
}
\details{
	The simulation starts by setting the dimensions of data matrix Ŷ (i.e. number of columns and rows). The number of columns was programmed to be equal to the number of expected species, while the number of rows (Nt) was defined arbitrarily as the potential number of sampling units (n) per site, multiplied by the potential number of sites. For each sampling unit, from 1 to n, at each site in Ŷ, the presence/absence of each species was simulated using random binomial values with the probability of occurrence set equal to that estimated from the pilot data. Exception to this rule were species that appear in all pilot samples, in which case the probability of occurrence was set as 0.9. For species with positive or negative spatial autocorrelation, the probability of occurrence was set as default at 0.9 or 0.1, respectively, each time it was simulated as present in the preceding sampling unit but restarting from the original probability at each new site. Once created, the P/A matrixes were converted to matrixes of abundances replacing presences by random values from an adequate statistical distribution and parameters were set equal to those estimated in the pilot data. Counts of individuals were generated using Poisson or negative binomial distributions, depending on the degree of aggregation of each species in the pilot data (McArdle & Anderson 2004; Anderson & Walsh 2013). When abundances were measured as a continuous variable (i.e. coverage, biomass), they were generated using the lognormal distribution. The simulation procedure was repeated to generate as many simulated data matrixes as needed.
}
\value{
\item{simulated.data}{
The function returns an object of class List, that includes all simulated data. This object will be used by \code{\link[SSP]{sampsd}} and \code{\link[SSP]{datquality}}}
}
\references{
Anderson, M. J., & Walsh, D. C. I. (2013). PERMANOVA, ANOSIM, and the Mantel test in the face of heterogeneous dispersions: What null hypothesis are you testing? Ecological Monographs, 83(4), 557-574. doi:10.1890/12-2010.1

McArdle, B. H., & Anderson, M. J. (2004). Variance heterogeneity, transformations, and models of species abundance: a cautionary tale. Canadian Journal of Fisheries and Aquatic Sciences, 61, 1294-1302.
}
\author{
Edlin Guerra-Castro (edlinguerra@gmail.com) and Maite Mascaro (mmm@ciencias.unam.mx)
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## micromollusk from 'Cayo Nuevo', Campeche bank, Gulf of Mexico, Yucatan, Mexico
data(micromollusk)
data(micromolluskGEO)

#Estimation of parameters of pilot data
dat.par<-assempar (data = micromollusk,
                    type= "P/A",
                    spatial.str = TRUE,
                    coords = micromolluskGEO,
                    moranI.alfa=0.05,
                    spatial.unsee.sp= "random")

#Simulation of 100 data sets, each one with 1000 potential sampling units from a single site
dat.sim<-simdata(dat.par, cases= 100, n= 1000, sites = 1)


##Sponges from Alacranes National Park, Gulf of Mexico
data(sponges)
data(spongesGEO)

#Estimation of parameters of pilot data
dat.par<-assempar (data = sponges,
                    type= "counts",
                    spatial.str = TRUE,
                    coords = spongesGEO,
                    moranI.alfa=0.05,
                    spatial.unsee.sp= "random")

#Simulation of 10 data sets, each one with 20 potential sampling units (transects) in 20 sites.
dat.sim<-SimData(dat.par, cases= 10, n= 20, sites = 20)
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ multivariate	&	Multivariate Techniques }% use one of  RShowDoc("KEYWORDS")
\keyword{ ecological communities }

