% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{SimData}
\alias{SimData}
\title{simulate assemblages}
\usage{
SimData(paramaters, cases, n, sites, positive.sp = 0.9,
  negative.sp = 0.1)
}
\arguments{
\item{paramaters}{list of parameters estimated by Assempar}

\item{cases}{number of data sets to be simulated}

\item{n}{total number of samples to simulate in each site}

\item{sites}{total number of sites to simulate in each data set}

\item{positive.sp}{probability of occurrence of a species once it has been
detected in a previos sample. The default is 0.9 for positive, 0.1 for
negative occurrences.}

\item{negative.sp}{probability of occurrence of a species once it has been
detected in a previos sample. The default is 0.9 for positive, 0.1 for
negative occurrences.}
}
\value{
The function returns an object of class List including all the simulated data sets to be used by DatQuality and SampSD.
}
\description{
SimData simulates matrixes (as many as
requested) using estimated parameters in the list generated by AssemPar.
}
\examples{
SimData(paramaters, cases, n, sites, positive.sp, negative.sp)

}
