\name{sponges}
\alias{sponges}
\docType{data}
\title{
Sponges in Alacranes Reef National Park (ARNP), Gulf of Mexico, Mexico
}
\description{
Counts of 41 species of sponges in 36 transects of 20 m * 1 m across 8 sites around ARNP}

\usage{data("sponges")}
\format{
  A data frame with 36 observations on the following 42 variables.
  \describe{
    \item{\code{site}}{Factor w/ 6 levels}
    \item{\code{Agelas.clathrodes}}{a numeric vector}
    \item{\code{Agelas.dispar}}{a numeric vector}
    \item{\code{Agelas.tubulata}}{a numeric vector}
    \item{\code{Agelas.wiedenmayeri}}{a numeric vector}
    \item{\code{Aiolocroia.crassa}}{a numeric vector}
    \item{\code{Amphimedon.copressa}}{a numeric vector}
    \item{\code{Aplysina.archeri}}{a numeric vector}
    \item{\code{Aplysina.cauliformis}}{a numeric vector}
    \item{\code{Aplysina.fistularis}}{a numeric vector}
    \item{\code{Aplysina.fulva}}{a numeric vector}
    \item{\code{Aplysina.insularis}}{a numeric vector}
    \item{\code{Aplysina.lacunosa}}{a numeric vector}
    \item{\code{Callyspongia.plicifera}}{a numeric vector}
    \item{\code{Callyspongia.vaginalis}}{a numeric vector}
    \item{\code{Callispongia.fallax}}{a numeric vector}
    \item{\code{Callispongia.armigera}}{a numeric vector}
    \item{\code{Cliona.delitrix}}{a numeric vector}
    \item{\code{Cliona.varians}}{a numeric vector}
    \item{\code{Cribochalina.vascolum}}{a numeric vector}
    \item{\code{Dragmacidon.sp.}}{a numeric vector}
    \item{\code{Dysidea.variabilis}}{a numeric vector}
    \item{\code{Ectyoplasia.ferox}}{a numeric vector}
    \item{\code{Geodia.neptuni}}{a numeric vector}
    \item{\code{Hymeniacidon.caerulea}}{a numeric vector}
    \item{\code{Iotrochota.birotulata}}{a numeric vector}
    \item{\code{Igernella.notabilis}}{a numeric vector}
    \item{\code{Ircinia.felix}}{a numeric vector}
    \item{\code{Ircinia.strobilina}}{a numeric vector}
    \item{\code{Monanchora.arbuscula}}{a numeric vector}
    \item{\code{Mycale.laxissima}}{a numeric vector}
    \item{\code{Mycale.laevis}}{a numeric vector}
    \item{\code{Nipahtes.amorpha}}{a numeric vector}
    \item{\code{Niphates.erecta}}{a numeric vector}
    \item{\code{Niphathes.digitalis}}{a numeric vector}
    \item{\code{Phorbas.amaranthus}}{a numeric vector}
    \item{\code{Scopalina.rutzleri}}{a numeric vector}
    \item{\code{Svenezea.flava}}{a numeric vector}
    \item{\code{Spirastrella.coccinea}}{a numeric vector}
    \item{\code{Verongula.reswigui}}{a numeric vector}
    \item{\code{Verongula.rigida}}{a numeric vector}
    \item{\code{Xestospongia.muta}}{a numeric vector}
  }
}
\details{
This data corresponds to a pilot study on sponge biodiversity in reef habitats in the Yucatán shelf (Ugalde et al., 2015)}

\source{
https://biotaxa.org/Zootaxa/article/view/zootaxa.3911.2.1
}
\references{
Ugalde, D., Gomez, P., & Simoes, N. (2015). Marine sponges (Porifera: Demospongiae) from the Gulf of Mexico, new records and redescription of Erylus trisphaerus (de Laubenfels, 1953). Zootaxa, 3911(2), 151-183.
}
\examples{
data(sponges)
str(sponges)
}
\keyword{datasets}
\keyword{SSP}
