\name{spongesGEO}
\alias{spongesGEO}
\docType{data}
\title{
Geographic coordinates of transects in \code{\link[SSP]{sponges}} data
}
\description{
Spatial coordinates of each transect
}
\usage{data("spongesGEO")}
\format{
  A data frame with 36 observations on the following 2 variables.
  \describe{
    \item{\code{Y}}{a numeric vector}
    \item{\code{X}}{a numeric vector}
  }
}
\details{
Geographic coordinates are given in decimal degrees
}
\source{
https://biotaxa.org/Zootaxa/article/view/zootaxa.3911.2.1
}
\references{
Ugalde, D., Gomez, P., & Simoes, N. (2015). Marine sponges (Porifera: Demospongiae) from the Gulf of Mexico, new records and redescription of Erylus trisphaerus (de Laubenfels, 1953). Zootaxa, 3911(2), 151-183.
}
\examples{
data(spongesGEO)

}
\keyword{datasets}
