\name{datquality}
\alias{datquality}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Diversity metrics of simulated and original data
}
\description{
The function estimates the average number of species and the Simpson diversity index per sampling unit, as well as the total multivariate dispersion of pilot data and simulated data
}

\usage{
datquality(data, dat.sim, Par, transformation, method)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{data}{
Data frame giving P/A or abundance (counts or coverage) of species
}
  \item{dat.sim}{
List of simulated data generated by simdata
}
  \item{Par}{
List of parameters generated by assempar
}
  \item{transformation}{
Mathematical function to reduce the weight of dominant species: 'square root', 'fourth root', 'Log (X+1)', 'P/A', 'none'
}
  \item{method}{
The appropiate distance/dissimilarity metric. The function \code{\link[vegan]{vegdist}} is call for that purpose
}
}
\details{
The quality of the simulated data sets was assessed by the resemblance these had to the pilot data considering the following estimations: (i) the average number of species per sampling unit, (ii) evenness, defined as the average Simpson diversity index per sampling unit, and (iii) the multivariate dispersion (MVD), measured as the average dissimilarity from all sampling units to the main centroid in the space of the dissimilarity measure used (Anderson 2006). Whereas it is desirable that (i) and (ii) are similar between simulated and pilot data, MVD is expected to be greater in the simulated matrixes. To assess the magnitude of this difference, 0.95 quantiles of the MVD were obtained for all simulated data sets.
}
\value{
\item{divmetrics}{
A data frame that includes the metrics for original and simulated data
}
}
\references{
Anderson, M.J. (2006) Distance-based tests for homogeneity of multivariate dispersions. Biometrics, 62, 245-253
}
\author{
Edlin Guerra-Castro (edlinguerra@gmail.com) and Maite Mascaro (mmm@ciencias.unam.mx)
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
## micromollusk from 'Cayo Nuevo', Campeche bank, Gulf of Mexico, Yucatan, Mexico
data(micromollusk)
data(micromolluskGEO)

#Estimation of parameters of pilot data
dat.par1<-assempar (data = micromollusk,
                    type= "P/A",
                    spatial.str = TRUE,
                    coords = micromolluskGEO,
                    moranI.alfa=0.05,
                    spatial.unsee.sp= "random")

#Simulation of 10 data sets, each one with 100 potential sampling units from a single site
dat.sim1<-simdata(dat.par1, cases= 10, n= 100, sites = 1)

#Estimation of diversity metrics of original and simulated data
dat.q1<-datquality(data = micromollusk,
                   dat.sim = dat.sim1,
                   Par = dat.par1,
                   transformation = "none",
                   method = "jaccard" )

##Sponges from Alacranes National Park, Gulf of Mexico
data(sponges)
data(spongesGEO)

#Estimation of parameters of pilot data
dat.par2<-assempar(data = sponges,
                    type= "counts",
                    spatial.str = TRUE,
                    coords = spongesGEO,
                    moranI.alfa=0.05,
                    spatial.unsee.sp= "random")

#Simulation of 10 data sets, each one with 20 potential sampling units (transects) in 20 sites.
dat.sim2<-simdata(dat.par2, cases= 10, n= 20, sites = 20)

#Estimation of diversity metrics of original and simulated data
dat.q2<-datquality(data = sponges,
                   dat.sim = dat.sim2,
                   Par = dat.par2,
                   transformation = "square root",
                   method = "bray")
}
