\name{sampsd}
\alias{sampsd}

\title{Sampling simulated data and estimates multivariate standard errors

}
\description{
Each set of simulated data is sampled many times for each sampling effort, from 2 replicates to those defined as an argument in the function. Then, distance-based multivariate standard errors is estimated using pseudo-variance (for single site evaluation) or Mean Squares estimates in a linear model (for multisite evaluation).
}
\usage{
sampsd(dat.sim, Par, transformation, method, multi.site = TRUE, n, p.n, sites, p.s, k)
}

\arguments{
  \item{dat.sim}{
A list of data sets generated by \code{\link{simdata}}
}
  \item{Par}{
A list of parameters estimated by \code{\link{assempar}}
}
  \item{transformation}{
Mathematical function to reduce the weight of very dominant species: 'square root', 'fourth root', 'Log (X+1)', 'P/A', 'none'
}
  \item{method}{
The appropiate distance/dissimilarity metric. The function \code{\link[vegan]{vegdist}} is call for that purpose.
  }
  \item{multi.site}{
Logical argument indicating if several sites were simulated
  }
  \item{n}{
Total number of samples simulated for each site
  }
  \item{p.n}{
Maxinum number of samples to take at each site. Can be equal or less than n
  }
  \item{sites}{
Total number of sites simulated for each data set
  }
  \item{p.s}{
Maxinum number of sites to sample at each data set
  }
  \item{k}{
Number of repetitions of each combination between n and sites
  }
}
\details{
Sampling without replacement is performed several times (e.g. 100) for each subset of samples of size n = 2, 3, 4 to N within each simulated matrix. If sites had been generated, subsets of sites of size m = 2, 3, 4 to M are first sampled followed by the sampling units for each site using inclusion probabilities and self-weighted two-stage sampling (Tille, 2011). Additionally, each combination of n and m are repeated several times (e.g. 100) for all simulated matrixes. Suitable pre-treatments of data matrixes are applied in each case and distance/similarity matrixes estimated using the appropriate coefficient. When simulations were done for a single site, the MultSE is estimated as \eqn{\sqrtV/n}, being V the pseudo variance measured at each sample of size n (Anderson & Santana-Garcon, 2015). When several sites were generated, MultSE are estimated using the residual mean squares and the sites mean squares from a PERMANOVA model (Anderson & Santana-Garcon, 2015).
}
\value{
\item{mse.results}{
A matrix including all estimated MultSE for each simulated data, combination of sample replicates and sites for each k repetition. This matrix will be used by \code{\link{summary_ssp}}}
}
\references{
Anderson, M.J. & Santana-Garcon, J. (2015) Measures of precision for dissimilarity-based multivariate analysis of ecological communities. Ecology Letters, 18, 66-73

Tille, Y. (2011) Sampling Algorithms. International Encyclopedia of Statistical Science (ed. M. Lovric), pp. 1273-1274. Springer Berlin Heidelberg, Berlin, Heidelberg
}
\author{
Edlin Guerra-Castro (edlinguerra@gmail.com), Juan Carlos Cajas, Juan Jose Cruz-Motta, Nuno Simoes and Maite Mascaro (mmm@ciencias.unam.mx).
}
\note{
For quick exploratory analyzes, keep the number of repetitions small. Once you have explored the behavior of the MultSE, you can repeat the process keeping k-value large (e.g. 100). This process will take some time, it will depend on the power of your computer.
}

\seealso{
\code{\link{assempar}}, \code{\link{simdata}}, \code{\link{summary_ssp}}, \code{\link[vegan]{vegdist}}}


\examples{
###To speed up the simulation of these examples, the cases, sites and n were set small.

##Single site: micromollusk from Cayo Nuevo (Yucatan, Mexico)
data(micromollusk)

#Estimation of parameters of pilot data
par.mic<-assempar (data = micromollusk,
                    type= "P/A",
                    Sest.method = "average")

#Simulation of 10 data sets, each one with 100 potential sampling units from a single site
sim.mic<-simdata(par.mic, cases= 10, n= 100, sites = 1)

#Sampling and estimation of MultSE for each sample size (few repetitions to speed up the example)
sam.mic<-sampsd(sim.mic,
               par.mic,
               transformation = "P/A",
               method = "jaccard",
               multi.site = FALSE,
               n=100,
               p.n = 50,
               sites = 1,
               p.s = 1,
               k=3)

##Multiple sites: Sponges from Alacranes National Park (Yucatan, Mexico).
data(sponges)

#Estimation of parameters of pilot data
par.spo<-assempar(data = sponges,
                  type= "counts",
                  Sest.method = "average")

#Simulation of 3 data sets, each one with 15 potential sampling units in 10 sites.
sim.spo<-simdata(par.spo, cases= 10, n= 15, sites = 10)

#Sampling and estimation of MultSE for each sampling design (few
#repetitions to speed up the example)

sam.spo<-sampsd(sim.spo,
                par.spo,
                transformation = "square root",
                method = "bray",
                multi.site = TRUE,
                n=15,
                p.n = 10,
                sites = 10,
                p.s = 10,
                k=3)

}

\keyword{ multivariate	&	Multivariate Techniques }
\keyword{ ecological communities }
\keyword{ numerical ecology }
